+++ packages/bodiless-layouts-ui/src/index.ts
@@ -13,4 +13,4 @@
  */
 
 // eslint-disable-next-line import/prefer-default-export
-export { default as FlexboxGrid } from './FlexboxGrid';
+export { default as FlowContainerGrid } from './FlowContainerGrid';