+++ packages/bodiless-layouts/src/FlowContainerGrid/StaticFlowContainer.tsx
@@ -15,32 +15,32 @@
 import React, { FC, PropsWithChildren } from 'react';
 import { withNode } from '@bodiless/core';
 import { useItemHandlers } from './helpers';
-import { StaticFlexboxProps, FlexboxItem } from './types';
+import { StaticFlowContainerProps, FlowContainerItem } from './types';
 
 const NodeProvider = withNode<PropsWithChildren<{}>, any>(React.Fragment);
 
-const StaticFlexbox: FC<StaticFlexboxProps> = ({ components }) => {
+const StaticFlowContainer: FC<StaticFlowContainerProps> = ({ components }) => {
   const items = useItemHandlers().getItems();
   return (
     // When in a static mode we don't want to use `bl-*` prefixed classes.
     <div className="flex flex-wrap">
       {items
-        .map((flexboxItem: FlexboxItem) => {
-          const ChildComponent = components[flexboxItem.type];
+        .map((flowContainerItem: FlowContainerItem) => {
+          const ChildComponent = components[flowContainerItem.type];
           // TODO: Inhance this notification when the data is bad
           if (!ChildComponent) {
-            throw new Error(`${flexboxItem.type} is not an allowed content type`);
+            throw new Error(`${flowContainerItem.type} is not an allowed content type`);
           }
           return (
             <div
-              key={`flex-${flexboxItem.uuid}`}
+              key={`flex-${flowContainerItem.uuid}`}
               className={
-                  (flexboxItem.wrapperProps
-                    && flexboxItem.wrapperProps.className)
+                  (flowContainerItem.wrapperProps
+                    && flowContainerItem.wrapperProps.className)
                   || ''
                 }
             >
-              <NodeProvider nodeKey={flexboxItem.uuid}>
+              <NodeProvider nodeKey={flowContainerItem.uuid}>
                 <ChildComponent />
               </NodeProvider>
             </div>
@@ -51,6 +51,6 @@ const StaticFlexbox: FC<StaticFlexboxProps> = ({ components }) => {
   );
 };
 
-StaticFlexbox.displayName = 'Flexbox';
+StaticFlowContainer.displayName = 'FlowContainer';
 
-export default StaticFlexbox;
+export default StaticFlowContainer;