+++ packages/bodiless-layouts/src/SortableResizableWrapper/index.tsx
@@ -15,7 +15,7 @@
 import { SortableElement, SortableHandle } from 'react-sortable-hoc';
 import React, { ComponentType, HTMLProps } from 'react';
 import CleanReresizable, { ResizeCallback, ResizableProps } from 're-resizable';
-import { SnapData } from '../FlexboxGrid/utils/appendTailwindWidthClass';
+import { SnapData } from '../FlowContainerGrid/utils/appendTailwindWidthClass';
 
 const ENABLED_DRAG_SIDES = {
   top: false,